sof: include: Move forward declarations to new <sof/types.h>
A small number of forward declarations are very specific and don't need
to be moved. The others were moved into a new header.

A future commit intends to use this header instead of others when only
type declarations (structures and typedefs) are needed from one of the
other headers.
