+++ src/include/sof/audio/eq_iir/eq_iir.h
@@ -10,10 +10,8 @@
 #ifndef __SOF_AUDIO_EQ_IIR_EQ_IIR_H__
 #define __SOF_AUDIO_EQ_IIR_EQ_IIR_H__
 
+#include <sof/types.h>
 #include <stdint.h>
-
-struct comp_buffer;
-struct comp_dev;
 
 /** \brief IIR EQ processing functions map item. */
 struct eq_iir_func_map {
