+++ src/include/sof/audio/eq_iir/iir.h
@@ -10,10 +10,9 @@
 #ifndef __SOF_AUDIO_EQ_IIR_IIR_H__
 #define __SOF_AUDIO_EQ_IIR_IIR_H__
 
+#include <sof/types.h>
 #include <stddef.h>
 #include <stdint.h>
-
-struct sof_eq_iir_header_df2t;
 
 /* Get platforms configuration */
 #include <config.h>
